moved adapter to src/helm/benchmark/adaptation
