+++ src/helm/benchmark/run_expander.py
@@ -18,9 +18,9 @@ from helm.proxy.models import (
     ABLATION_MODEL_TAG,
 )
 from .runner import RunSpec
+from .adaptation.adapter import Substitution
 from .augmentations.perturbation import PerturbationSpec
 from .augmentations.data_augmenter import DataAugmenterSpec
-from helm.benchmark.adapter import Substitution
 
 
 class RunExpander(ABC):
